Rename lcd_i2c.c to lcd_i2c-update.c
I had the need to use this code with a four line display. As the author pointed out in the lcd_i2c_gotoxy function, I found that the software would not work for lines 3 and 4 of a four line display. In order to resolve this issue, I made a simple change to this function, which is within the lcd_i2c.c file; no other changes were necessary. Hopefully, this helps someone out!

Thanks,
Marshall